pyproject.toml
[tool.poetry]
name = "docx-translator"
version = "0.1.0"
description = ""
authors = ["Pandego <7780875+pandego@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
deep-translator = "^1.11.4"
ipykernel = "^6.29.5"
python-docx = "^1.1.2"
rich = "^13.7.1"
tqdm = "^4.66.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
